Remove DefinedSynthetic.
With this we have a single section hierarchy. It is a bit less code,
but the main advantage will be in a future patch being able to handle

foo = symbol_in_obj;

in a linker script. Currently that fails since we try to find the
output section of symbol_in_obj.  With this we should be able to just
return an InputSection from the expression.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@297313 91177308-0d34-0410-b5e6-96231b3b80d8