Fix wrong assertion failure.
Previously, if you have foo=bar in a definition file, this assertion
could fire because when symbols are read from file they could be mangled.
It seems that due to historical reasons underscore mangling scheme is
really ad-hoc, and I cannot find a clean way to handle this. I had
to just de-mangle symbols to search again.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@297357 91177308-0d34-0410-b5e6-96231b3b80d8