[ELF] - Move members of LinkerScript to LinkerScriptBase. NFC.
That moves all members that s possible to move for now (all which
does not depend on ELFT templating).

After that change LinkerScript contains only 8 methods in total,
and I believe it is possible to move them all after tweaking other
parts of linker. And we will be able to have single class for 
linkerscript at the end.


git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@297735 91177308-0d34-0410-b5e6-96231b3b80d8