[ELF] Add a NOP instruction to make LLVM generate a mapping symbol
This is a test change applicable for ARM targets.

Previously LLVM would generate a mapping symbol when code contains only
data. This was changed as part of

https://reviews.llvm.org/D30724

and to be more close to the ARM ELF ABI.

The test case is changed to check for behavior with minimal changes.

Patch by Shankar Easwaran

Differential Revision: https://reviews.llvm.org/D31500



git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@299212 91177308-0d34-0410-b5e6-96231b3b80d8