Don't resolve hidden undef to a DSO.
The ELF spec says:

all of the non-default visibility attributes, when applied to a symbol
reference, imply that a definition to satisfy that reference must be
provided within the current executable or shared object.

But we were trying to resolve those undef references to shared
symbols. That causes odd results like creating a got entry with
a relocation pointing to 0.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@299464 91177308-0d34-0410-b5e6-96231b3b80d8