[LLD][LinkerScript] Add support for segment NONE.
This patch adds support for segment NONE in linker scripts which enables the
specification that a section should not be assigned to any segment.

Note that GNU ld does not disallow the definition of a segment named NONE, which
if defined, effectively overrides the behaviour described above. This feature
has been copied.

Differential Revision: https://reviews.llvm.org/D34203

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@305700 91177308-0d34-0410-b5e6-96231b3b80d8