Keep the original symbol name when renamed.
Previously, when symbol A is renamed B, both A and B end up having
the same name. This is because name is a symbol's attribute, and
we memcpy symbols for symbol renaming.

This pathc saves the original symbol name and restore it after memcpy
to keep the original name.

This patch shouldn't change program's meaning, but names in symbol
tables make more sense than before.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@306036 91177308-0d34-0410-b5e6-96231b3b80d8