Merge OutputSectionCommand and OutputSection.
This is a bit of a hack, but it is *so* convenient.

Now that we create synthetic linker scripts when none is provided, we
always have to handle paired OutputSection and OutputsectionCommand and
keep a mapping from one to the other.

This patch simplifies things by merging them and creating what used to
be OutputSectionCommands really early.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@309311 91177308-0d34-0410-b5e6-96231b3b80d8