Add a special case for trivial alignment.
Normally to find the offset of a value in a section, we have to
compute the value since the alignment is defined on the final address.

If the alignment is trivial, we can skip the value computation. This
allows us to know the offset even in cases where we cannot yet know
the value.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@313777 91177308-0d34-0410-b5e6-96231b3b80d8