Merge addInputSec with OutputSection::addSection.
Previously, when we added an input section to an output section, we
called `OutputSectionFactory::addInputSec`. This isn't a good design
because, a factory class is intended to create a new object and
return it, but in this use case, it will never create a new object.
This patch fixes the design flaw.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@315138 91177308-0d34-0410-b5e6-96231b3b80d8