Don't create a dummy __tls_get_addr.
We just don't need one with the current setup.

We only error on undefined references that are used by some
relocation.

If we managed to relax all uses of __tls_get_addr, no relocation uses
it and we don't produce an error.

This is less code and fixes the case were we fail to relax. Before we
would produce a broken output, but now we produce an error.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@315334 91177308-0d34-0410-b5e6-96231b3b80d8