Handle MIPS-specific addend rules in computeAddend().
This patch merges computeAddend and computeMipsAddend.

Getting an addend for a relocation is usually pretty easy:
it is either in the r_addend field (if RELA) or in a target
section (if REL).

However, MIPS has many special rules that are different from
other ELF ABIs. I don't think there were technical reasons to
be different, but the reality is that they are different.
It is unfortunate that we had to pass many parameters to
computeAddend, but it seems unavoidable because of MIPS.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@315617 91177308-0d34-0410-b5e6-96231b3b80d8