Slightly simplify code and add comment.
This is not a mechanical transformation. Even though I believe this
patch is correct, I'm not 100% sure if lld with this patch behaves
exactly the same way as before on all edge cases. At least all tests
still pass.

I'm submitting this patch because it took almost a day to understand
this function, and I don't want to lose it.

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@315658 91177308-0d34-0410-b5e6-96231b3b80d8