Fix SizeOfImage in the PE header.
IIUC, SizeOfImage is the distance from the end of the last section to
the image base, rounded up to the page size. So the previous code is
wrong.

Should fix https://bugs.llvm.org/show_bug.cgi?id=34949

(It is nice to know that lld is already being used to create Putty
distribution binaries.)

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@316626 91177308-0d34-0410-b5e6-96231b3b80d8